+++ example/lib/samples/widget_extension_view.dart
@@ -5,7 +5,7 @@ class WidgetExtensionViews extends StatelessWidget {
   const WidgetExtensionViews({super.key});
 
   bool get isDataValid => false;
-  
+
   @override
   Widget build(BuildContext context) {
     return const Text('is data not found').toVisible(isDataValid);