+++ lib/src/utility/bundle/bundle_decoder.dart
@@ -6,12 +6,14 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:kartal/src/utility/bundle/i_network_model.dart';
 
-typedef ComputeCallback<T extends INetworkModel<T>, R> = Future<R> Function(_ComputeArgument<T>);
+typedef ComputeCallback<T extends INetworkModel<T>, R> = Future<R> Function(
+    _ComputeArgument<T>);
 
 class BundleDecoder extends _BaseBundleDecoder with _BundleHelpers {
   BundleDecoder(super.assetPath);
 
-  Future<R?> crackBundle<T extends INetworkModel<T>, R>({required T model}) async {
+  Future<R?> crackBundle<T extends INetworkModel<T>, R>(
+      {required T model}) async {
     try {
       final decodedData = await loadBundle<T, R>(super.assetPath, model: model);
       return decodedData;
@@ -22,11 +24,13 @@ class BundleDecoder extends _BaseBundleDecoder with _BundleHelpers {
 }
 
 mixin _BundleHelpers {
-  Future<R?> loadBundle<T extends INetworkModel<T>, R>(String path, {required T model}) async {
+  Future<R?> loadBundle<T extends INetworkModel<T>, R>(String path,
+      {required T model}) async {
     try {
       final bundle = await rootBundle.loadString(path);
 
-      final ComputeCallback<T, R> callBack = (_ComputeArgument<T> argument) => parse<T, R>(argument);
+      final ComputeCallback<T, R> callBack =
+          (_ComputeArgument<T> argument) => parse<T, R>(argument);
 
       return await compute(
         callBack,
@@ -40,13 +44,16 @@ mixin _BundleHelpers {
     }
   }
 
-  Future<R> parse<T extends INetworkModel<T>, R>(_ComputeArgument<T> argument) async {
+  Future<R> parse<T extends INetworkModel<T>, R>(
+      _ComputeArgument<T> argument) async {
     if (R == List<T>) {
       try {
-        final transformedData = (json.decode(argument.bundle) as List<dynamic>) //
-            .map((data) => argument.model.fromJson(data as Map<String, dynamic>)) //
-            .cast<T>()
-            .toList() as R;
+        final transformedData =
+            (json.decode(argument.bundle) as List<dynamic>) //
+                .map((data) =>
+                    argument.model.fromJson(data as Map<String, dynamic>)) //
+                .cast<T>()
+                .toList() as R;
 
         return transformedData;
       } catch (e) {